requirements.txt
boto3==1.21.31
botocore==1.24.46
Flask==2.0.3
Flask_Cors==3.0.9
numpy==1.23.4
orjson==3.8.1
Pillow==9.3.0
pytesseract==0.3.10
python-dotenv==0.21.0
PyYAML==6.0
torch==1.13.0
torchvision==0.14
Werkzeug==2.0.3
pdf2image==1.16.2